+++ riverpie/lib/riverpie.dart
@@ -18,6 +18,8 @@ export 'package:riverpie/src/notifier/types/view_provider_notifier.dart';
 export 'package:riverpie/src/observer/event.dart';
 export 'package:riverpie/src/observer/history_observer.dart';
 export 'package:riverpie/src/observer/observer.dart';
+export 'package:riverpie/src/observer/tracing_observer.dart'
+    show RiverpieTracingObserver;
 export 'package:riverpie/src/provider/types/async_notifier_provider.dart';
 export 'package:riverpie/src/provider/types/change_notifier_provider.dart';
 export 'package:riverpie/src/provider/types/future_family_provider.dart';