+++ riverpie_flutter/lib/riverpie_flutter.dart
@@ -3,4 +3,5 @@ export 'package:riverpie_flutter/src/consumer.dart';
 export 'package:riverpie_flutter/src/extension.dart';
 export 'package:riverpie_flutter/src/mixin.dart';
 export 'package:riverpie_flutter/src/scope.dart';
+export 'package:riverpie_flutter/src/tracing_page.dart';
 export 'package:riverpie_flutter/src/widget_rebuildable.dart';